# perf: ⚡️ inline function