inline scanner function that call multiple time, reduce function call overhead
master branch: 
```
test citm_catalog_json_large_ast   ... bench:  74,397,750 ns/iter (+/- 15,422,789)
test citm_catalog_json_large_value ... bench: 110,817,590 ns/iter (+/- 4,172,933)
test package_json_ast              ... bench:      81,854 ns/iter (+/- 1,644)
test package_json_value            ... bench:     130,190 ns/iter (+/- 1,455)
test tsconfig_json_ast             ... bench:      26,818 ns/iter (+/- 395)
test tsconfig_json_value           ... bench:      29,314 ns/iter (+/- 324)
```
perf/inline branch: 
```
test citm_catalog_json_large_ast   ... bench:  72,379,470 ns/iter (+/- 18,073,936)
test citm_catalog_json_large_value ... bench: 106,483,950 ns/iter (+/- 5,451,741)
test package_json_ast              ... bench:      76,892 ns/iter (+/- 1,476)
test package_json_value            ... bench:     121,578 ns/iter (+/- 4,517)
test tsconfig_json_ast             ... bench:      28,673 ns/iter (+/- 530)
test tsconfig_json_value           ... bench:      26,847 ns/iter (+/- 409)
```