@dsherret 2021-04-27
@IWANABETHATGUY is this actually faster? I would have thought the compiler would be smart enough to automatically inline these.