@dsherret 2021-05-09
I've found that the CI bench is not a good controlled environment and there is a lot of variation between runs. That said, I ran this on my computer and I don't see any improvement :(